Compiling 1_BACK.C
