+++ src/utility.h
@@ -2,5 +2,6 @@
 
 #include <stdint.h>
 
+void delay(int milliseconds);
 int8_t parse_decimal_int8_t(char* str);
 int16_t parse_decimal_int16_t(char* str);
